page file structure refactor
