+++ src/pages/index.tsx
@@ -1,8 +1,8 @@
 import ChangePassword from "./ChangePassword";
-import Checks from "./Checks";
+import Checks from "./Admin/Checks";
 import Home from "./Home";
 import Login from "./Login";
 import Me from "./Me";
-import Users from "./Users";
+import Users from "./Admin/Users";
 
 export { ChangePassword, Checks, Home, Login, Me, Users };
